+++ docs/Changelog.md
@@ -1,11 +1,16 @@
 # Changelog
 
-## v3.2.4
+## v3.3.0
+* :core: Create debounce methods
+* :searchview: [Breaking] remove reactive dependencies and stick with basic callbacks
+
+## v3.2.5
 * :core: Fix FAQ background
 * :core: Create FileUtils
 * :core: Create NotificationUtils
 * :core: Update swipe to remove most exceptions
 * :core: Make logging class functions inline
+* :core: Create removeIf for mutableIteratables
 * :core-ui: Move reactive libs to :searchview:
 
 ## v3.2.3